Added support for large files (2GB+) Rebuild Database button now hides if password is not set Fixed broken bugged gallery after page refresh Fixed sorting and rating buttons not having text in mobile/chrome browsers.

+++ server/helpers/dbImport/setup.sql
@@ -59,7 +59,7 @@ DROP TABLE IF EXISTS posts CASCADE;
 CREATE TABLE posts (
   id INTEGER PRIMARY KEY,
   hash BYTEA NOT NULL,
-  size INTEGER,
+  size BIGINT,
   width INTEGER,
   height INTEGER,
   duration FLOAT,
